added authentication and autherization for deleting post

+++ controllers/api/v1/post.js
@@ -21,17 +21,18 @@ module.exports.destroy = async (req, res) => {
     try {
         const {id} = req.params
         const post = await Post.findById(id)
-        post.remove()
-        console.log('post deleted')
-        await Comment.deleteMany({post: id})
-        console.log('comments on post also deleted')
-        return res.status(200).json({
-            message: 'post and associated comments deleted successfully'
-        })
+        if(post && post.user.toString() === req.user.id) {
+            post.remove()
+            console.log('post deleted')
+            await Comment.deleteMany({post: id})
+            console.log('comments on post also deleted')
+            return res.status(200).json({message: 'post deleted successfully'})
+        } else {
+            console.log('post not found or user id does not match')
+            return res.status(401).json({message: 'post not found or user id does not match'})
+        }
     } catch(err) {
         console.log('error in deleting post ', err)
-        return res.status(500).json({
-            message: 'Internal Server Error'
-        })
+        return res.status(401).json({message: 'ERROR'})
     }
 }
